+++ packages/rev-ui/src/actions/__tests__/SaveAction.tests.tsx
@@ -10,6 +10,7 @@ import { DetailView, IDetailViewContextProp } from '../../views/DetailView';
 import { SaveAction, ISaveActionProps } from '../SaveAction';
 import { withDetailViewContext } from '../../views/withDetailViewContext';
 import { IActionComponentProps } from '../types';
+import { ModelValidationResult } from 'rev-models/lib/validation/validationresult';
 
 describe('SaveAction', () => {
 
@@ -226,9 +227,27 @@ describe('SaveAction', () => {
             expect(receivedProps.detailViewContext.loadState).to.equal('SAVING');
         });
 
+        it('resets loadState and calls onError() with a validation error if validation fails', async () => {
+            expect(receivedProps.detailViewContext.loadState).to.equal('NONE');
+
+            const validResult = new ModelValidationResult(false);
+            receivedProps.detailViewContext.validate = (): any => Promise.resolve(validResult);
+
+            await receivedProps.doAction();
+
+            expect(onErrorCallback.callCount).to.equal(1);
+            const err = onErrorCallback.getCall(0).args[0];
+            expect(err).to.be.instanceof(rev.ValidationError);
+            expect(err.validation).to.equal(validResult);
+
+            expect(receivedProps.detailViewContext.loadState).to.equal('NONE');
+        });
+
         it('resets loadState and calls onError() if an error is returned from save()', async () => {
             expect(receivedProps.detailViewContext.loadState).to.equal('NONE');
 
+            const validResult = new ModelValidationResult(true);
+            receivedProps.detailViewContext.validate = (): any => Promise.resolve(validResult);
             const expectedError = new Error('ack!!!');
             receivedProps.detailViewContext.save = () => Promise.reject(expectedError);
 
@@ -244,6 +263,8 @@ describe('SaveAction', () => {
         it('resets loadState and calls onSuccess() if the save() method is successful', async () => {
             expect(receivedProps.detailViewContext.loadState).to.equal('NONE');
 
+            const validResult = new ModelValidationResult(true);
+            receivedProps.detailViewContext.validate = (): any => Promise.resolve(validResult);
             const expectedResult = 'Yay!';
             receivedProps.detailViewContext.save = () => Promise.resolve(expectedResult) as any;